Moving release steps into it's own module and removing unused imports in setup.
